+++ packages/sdk/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @reservoir0x/relay-sdk
 
+## 1.8.1
+
+### Patch Changes
+
+- e5d8e39: Sync api types
+
+## 1.8.0
+
+### Minor Changes
+
+- d430ab3: Overhaul analytics events
+
 ## 1.7.4
 
 ### Patch Changes